Update in the Api-file to fetch all movies
